requirements.txt
websockets==12.0
aiofiles==23.2.1
PyJWT==2.10.1
aiosqlite==0.19.0

# Testing
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
